requirements.txt
numpy
pandas
scikit-learn
opencv-python
matplotlib
Pillow
tensorflow
joblib
flask